moves exam/* components into feature specific packages
