+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/AssessmentItemRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam;
 
 import org.opentestsystem.rdw.reporting.exam.AssessmentItem;
