+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/JdbcAssessmentItemRepository.java
@@ -1,8 +1,6 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam;
 
 import com.google.common.collect.ImmutableList;
-import org.opentestsystem.rdw.reporting.exam.AssessmentItem;
-import org.opentestsystem.rdw.reporting.exam.AssessmentItemDifficultyRange;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.EmptyResultDataAccessException;
