+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/DefaultEthnicityService.java
@@ -1,7 +1,6 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search;
 
 import com.google.common.collect.ImmutableSet;
-import org.opentestsystem.rdw.reporting.exam.repository.EthnicityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
