+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/DefaultExamFilterOptionService.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
