+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/EthnicityRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam.search;
 
 import java.util.List;
