+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/ExamFilterOptionController.java
@@ -1,7 +1,7 @@
-package org.opentestsystem.rdw.reporting.web;
+package org.opentestsystem.rdw.reporting.exam.search;
 
-import org.opentestsystem.rdw.reporting.exam.ExamFilterOptionService;
-import org.opentestsystem.rdw.reporting.exam.ExamFilterOptions;
+import org.opentestsystem.rdw.reporting.exam.search.ExamFilterOptionService;
+import org.opentestsystem.rdw.reporting.exam.search.ExamFilterOptions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
