+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/ExamFilterOptionService.java
@@ -1,4 +1,6 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search;
+
+import org.opentestsystem.rdw.reporting.exam.search.ExamFilterOptions;
 
 /**
  * Responsible for managing exam filter options
