+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/SchoolYearService.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search;
 
 import java.util.Set;
