+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/StubSchoolYearService.java
@@ -1,6 +1,7 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search;
 
 import com.google.common.collect.ImmutableSet;
+import org.opentestsystem.rdw.reporting.exam.search.SchoolYearService;
 import org.springframework.stereotype.Service;
 
 import java.util.Set;
