+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/DefaultGroupExamItemService.java
@@ -1,7 +1,8 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
-import org.opentestsystem.rdw.reporting.exam.repository.AssessmentItemRepository;
-import org.opentestsystem.rdw.reporting.exam.repository.GroupExamItemRepository;
+import org.opentestsystem.rdw.reporting.exam.AssessmentItem;
+import org.opentestsystem.rdw.reporting.exam.AssessmentItemRepository;
+import org.opentestsystem.rdw.reporting.exam.ExamItemScore;
 import org.opentestsystem.rdw.reporting.security.GroupSecurityContext;
 import org.opentestsystem.rdw.reporting.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
