+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/DefaultGroupExamService.java
@@ -1,7 +1,9 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
 import com.google.common.collect.ImmutableList;
-import org.opentestsystem.rdw.reporting.exam.repository.GroupExamRepository;
+import org.opentestsystem.rdw.reporting.exam.Assessment;
+import org.opentestsystem.rdw.reporting.exam.Exam;
+import org.opentestsystem.rdw.reporting.exam.GroupExam;
 import org.opentestsystem.rdw.reporting.security.GroupSecurityContext;
 import org.opentestsystem.rdw.reporting.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
