+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/GroupExamController.java
@@ -1,7 +1,7 @@
-package org.opentestsystem.rdw.reporting.web;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
 import org.opentestsystem.rdw.reporting.exam.Assessment;
-import org.opentestsystem.rdw.reporting.exam.GroupExamService;
+import org.opentestsystem.rdw.reporting.exam.search.group.GroupExamService;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.GroupExam;
 import org.opentestsystem.rdw.reporting.security.User;
