+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/GroupExamItemController.java
@@ -1,7 +1,7 @@
-package org.opentestsystem.rdw.reporting.web;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
-import org.opentestsystem.rdw.reporting.exam.GroupExamItemScores;
-import org.opentestsystem.rdw.reporting.exam.GroupExamItemService;
+import org.opentestsystem.rdw.reporting.exam.search.group.GroupExamItemScores;
+import org.opentestsystem.rdw.reporting.exam.search.group.GroupExamItemService;
 import org.opentestsystem.rdw.reporting.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
