+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/GroupExamItemRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
 import org.opentestsystem.rdw.reporting.exam.ExamItemScore;
 import org.opentestsystem.rdw.reporting.security.GroupSecurityContext;
