+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/GroupExamItemScores.java
@@ -1,4 +1,7 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search.group;
+
+import org.opentestsystem.rdw.reporting.exam.AssessmentItem;
+import org.opentestsystem.rdw.reporting.exam.ExamItemScore;
 
 import javax.validation.constraints.NotNull;
 import java.util.List;
