+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/GroupExamItemService.java
@@ -1,5 +1,6 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
+import org.opentestsystem.rdw.reporting.exam.search.group.GroupExamItemScores;
 import org.opentestsystem.rdw.reporting.security.User;
 import org.springframework.security.access.prepost.PreAuthorize;
