+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/GroupExamService.java
@@ -1,5 +1,6 @@
-package org.opentestsystem.rdw.reporting.exam;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
+import org.opentestsystem.rdw.reporting.exam.Assessment;
 import org.opentestsystem.rdw.reporting.exam.Exam;
 import org.opentestsystem.rdw.reporting.exam.GroupExam;
 import org.opentestsystem.rdw.reporting.security.User;
