+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/exam/search/group/JdbcGroupExamRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
