+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/web/SchoolYearController.java
@@ -1,6 +1,7 @@
 package org.opentestsystem.rdw.reporting.web;
 
-import org.opentestsystem.rdw.reporting.exam.SchoolYearService;
+import org.opentestsystem.rdw.reporting.exam.search.ExamFilterOptionController;
+import org.opentestsystem.rdw.reporting.exam.search.SchoolYearService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
