+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/exam/JdbcAssessmentItemRepositoryIT.java
@@ -1,11 +1,9 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.reporting.ITDataSourceConfiguration;
 import org.opentestsystem.rdw.reporting.RepositoryIT;
-import org.opentestsystem.rdw.reporting.exam.AssessmentItem;
-import org.opentestsystem.rdw.reporting.exam.AssessmentItemDifficulty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
