+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/exam/search/group/JdbcGroupExamItemRepositoryIT.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
