+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/exam/search/group/JdbcGroupExamRepositoryIT.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.repository;
+package org.opentestsystem.rdw.reporting.exam.search.group;
 
 import com.google.common.collect.ImmutableList;
 import org.junit.Test;
@@ -20,7 +20,7 @@ import static org.opentestsystem.rdw.reporting.security.GroupSecurityContexts.st
 
 @RunWith(SpringRunner.class)
 @RepositoryIT
-@ContextConfiguration(classes = {ITDataSourceConfiguration.class, JdbcGroupExamRepository.class, JdbcEthnicityRepository.class})
+@ContextConfiguration(classes = {ITDataSourceConfiguration.class, JdbcGroupExamRepository.class})
 @Sql(statements = {
         "INSERT INTO gender (id, code) VALUES (-1, 'g1');\n",
         "INSERT INTO district (id, natural_id, name) VALUES (-1, 'district1', 'district1');\n",
